+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 All notable changes to this project will be documented in this file.
 ## [5.8.0] - 2018-08-14
 ### Added
+- Bow enhancement.
 ### Fixed
+- Fixed an error that prevents gui from opening.
 ### Changed
+- Renamed `enableFireworkDamage: false` to `enablePreventFireworkDamage: true`.
 ### Removed
 
 ## [5.7.0] - 2018-08-13